refractors main.py
